requirements.txt
flask
python-dotenv
Flask-Mail
slackclient
validate_email